+++ swr/src/main/kotlin/dev/burnoo/compose/swr/retry/SWROnRetry.kt
@@ -3,5 +3,5 @@ package dev.burnoo.compose.swr.retry
 import dev.burnoo.compose.swr.config.SWRConfig
 
 typealias SWROnRetry<K, D> = suspend (
-    error: Throwable, key: K, config: SWRConfig<K, D>, attempt: Int
+    error: Throwable, key: K, config: SWRConfig<K, D>, attempt: Long
 ) -> Boolean
